# swagger/swagger.yaml
swagger: '2.1'
info:
  title: Docker Vulnerability Manager API
  description: Docker containers' source code freshness management API
  version: 0.19.0
consumes:
  - application/json
produces:
  - application/json
basePath: /api
components:
  parameters:
    token:
      in: header
      name: token
      required: true
      schema:
        type: string
      description: authentication token
      example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6IjVhOTVhNzBiYjljMjQ1MzA2YjE2YzE0NSIsImlhdCI6MTUxOTkyMzc5NiwiZXhwIjoxNTE5OTI3Mzk2fQ.e5cKEEGtDb54AVlnZHf-P132x1LTrS4WrgXywfmkSkk
...
    imageName:
      in: body
      name: imageName
      required: true
      schema:
        type: string
      description: name of the Docker image
      example: pawelpaszki/vuln-demo-2-node
...
    checkOnly:
      in: body
      name: checkOnly
      required: false
      schema:
        type: boolean
      descrpiton: provided if persistence of the vulnerability check is not required
paths:
...
  /imagefreshness:
    put:
      summary: save vulnerability check
      description: scan container's source code and record found vulnerabilities
      operationId: performVulnerabilityCheck
      tags:
        - ImageFreshness
      parameters:
        - $ref: '#/components/parameters/token'
        - $ref: '#/components/parameters/imageName'
        - $ref: '#/components/parameters/checkOnly'
      responses:
        '201':
          description: OK
          schema:
            type: object
            properties:
              message:
                type: string
                example: Vulnerability check persisted successfully
              updates:
                type: object
                example: []
              vulnerabilityCheckRecord:
                type: object
                example:  []

        '403':
          description: Unable to persist vulnerability check. Docker image's name required
          schema:
            type: object
            properties:
              error:
                type: string
                example: Unable to persist vulnerability check. Docker image's name required
        '404':
          description: Source code not extracted for this image
          schema:
            type: object
            properties:
              error:
                type: string
                example: Source code not extracted for this image
        '409':
          description: Vulnerability check already persisted for today's date
          schema:
            type: object
            properties:
              error:
                type: string
                example: Vulnerability check already persisted for today's date
        '500':
          description: Unable to persist vulnerability check
          schema:
            type: object
            properties:
              error:
                type: string
                example: Unable to persist vulnerability check
    post:
      summary: Create an image freshness entry
      description: create new image freshness entry for a Docker image
      operationId: createImageFreshness
      tags:
        - ImageFreshness
      parameters:
        - $ref: '#/components/parameters/token'
        - $ref: '#/components/parameters/imageName'
      responses:
        '201':
          description: OK
          schema:
            type: object
            properties:
              message:
                type: string
                example: Image freshness created saved successfully
        '403':
          description: Unable to create image freshness entry
          schema:
            type: object
            properties:
              message:
                type: string
                example: Unable to create image freshness entry
    delete:
      summary: Delete image freshness entries
      description: Delete image freshness entries
      operationId: deleteImageFreshnessEntryies
      parameters:
        - $ref: '#/components/parameters/token'
      tags:
        - ImageFreshness
      responses:
        '200':
          description: OK
          schema:
            type: object
            properties:
              message:
                type: string
                example: Image freshness entries deleted successfully
    get:
      summary: List all imagefreshness entries
      description:  Lists all of the imagefreshness entries with the recorded vulnerabilities
      operationId: listImageFreshnessEntries
      parameters:
        - $ref: '#/components/parameters/token'
      tags:
        - ImageFreshness
      responses:
        '200':
          description: OK
          schema:
            type: object
            properties:
              name:
                type: string
                example: pawelpaszki/vuln-demo-2-node
              lowVulnCount:
                type: number
                example: 0
              mediumVulnCount:
                type: number
                example: 1
              highVulnCount:
                type: number
                example: 1
              vulnerabilityCheckRecords:
                type: object
                example: [{
                  "date": "2018-02-15T12:46:47.872Z",
                  "lowSeverity": [],
                  "mediumSeverity": [
                    {
                      "dependencyPath": "marked@0.3.7",
                      "description": "Cross-site Scripting (XSS)",
                      "name": "marked@0.3.7",
                      "remediation": "marked@0.3.7 to marked@0.3.9"
                    },
                    {
                      "dependencyPath": "mqtt@2.0.1",
                      "description": "Denial of Service (DoS)",
                      "name": "mqtt@2.0.1",
                      "remediation": "mqtt@2.0.1 to mqtt@2.15.0"
                    }
                  ],
                  "highSeverity": [
                    {
                      "dependencyPath": "marked@0.3.7",
                      "description": "Cross-site Scripting (XSS)",
                      "name": "marked@0.3.7",
                      "remediation": "marked@0.3.7 to marked@0.3.9"
                    },
                    {
                      "dependencyPath": "marked@0.3.7",
                      "description": "Regular Expression Denial of Service (ReDoS)",
                      "name": "marked@0.3.7",
                      "remediation": "marked@0.3.7 to marked@0.3.9"
                    }
                  ]
                }]
...